Set up for mutliple pages in the app.
React router allows for different "pages" but all pages download the
whole bundle. Original page is Home and Projects is just a template for
now.
Resume was added as a binary to the public folder so that it can be
linked to by name.
Bootstrap was being linked multiple times. This was fixed by only
importing scss that compiles to empty css (functions/mixins) in the
component scss. Also needed to reorder imports so bootstrap came before
them.